requirements.txt
fastapi
uvicorn
twilio
azure-cognitiveservices-speech
python-multipart
